Finished traps
I added the functionality to disarm traps and specify the items that work for disarming a trap. However it is not yet usable because I don't know how to use regex. But the disarmAction method in MessageParser should work. I also added a parseDirection method in MessageParser because duplicate code was used in moveAction and disarmAction.